fix: Make Payload < 10Mo limit
For user batch import Brevo specified that "Maximum allowed json body size is 10MB".
Our instance have grown to quickly so we reach that limit.
The proposed solution is to set a limit of 5000 user per batch (as it
was working when our instance had this volume of users).

Other improvement, the script can now give another exit status than 0.
That must ensure that our monitoring can detect when a Cron fails.